Switch to SolidJS
Qwik just doesn't seem well-suited for a rich SPA experience. It's too
difficult to keep data cached and components in memory. I also find it
hard to figure out what will be run on the server vs on the client,
which I think is intentional - but not ideal for my use-case.